Refactor testing utilities for Event shims
