+++ packages/@react-aria/slider/test/useSliderThumb.test.js
@@ -1,4 +1,5 @@
 import {fireEvent, render, screen} from '@testing-library/react';
+import {installMouseEvent} from '@react-spectrum/test-utils';
 import * as React from 'react';
 import {renderHook} from '@testing-library/react-hooks';
 import {useRef} from 'react';
@@ -103,6 +104,8 @@ describe('useSliderThumb', () => {
       widthStub.mockReset();
     });
 
+    installMouseEvent();
+
     let stateRef = React.createRef();
 
     function RangeExample(props) {
@@ -202,6 +205,7 @@ describe('useSliderThumb', () => {
     afterAll(() => {
       widthStub.mockReset();
     });
+    installMouseEvent();
 
     let stateRef = React.createRef();
