+++ packages/@react-spectrum/test-utils/src/events.ts
@@ -29,6 +29,61 @@ export function triggerPress(element, opts = {}) {
 }
 
 /**
+ * Enables reading pageX/pageY from fireEvent.mouse*(..., {pageX: ..., pageY: ...}).
+ */
+export function installMouseEvent() {
+  beforeAll(() => {
+    let oldMouseEvent = MouseEvent;
+    // @ts-ignore
+    global.MouseEvent = class FakeMouseEvent extends MouseEvent {
+      _init: {pageX: number, pageY: number};
+      constructor(name, init) {
+        super(name, init);
+        this._init = init;
+      }
+      get pageX() {
+        return this._init.pageX;
+      }
+      get pageY() {
+        return this._init.pageY;
+      }
+    };
+    // @ts-ignore
+    global.MouseEvent.oldMouseEvent = oldMouseEvent;
+  });
+  afterAll(() => {
+    // @ts-ignore
+    global.MouseEvent = global.MouseEvent.oldMouseEvent;
+  });
+}
+
+export function installPointerEvent() {
+  beforeAll(() => {
+    // @ts-ignore
+    global.PointerEvent = class FakePointerEvent extends MouseEvent {
+      _init: {pageX: number, pageY: number, pointerType: string};
+      constructor(name, init) {
+        super(name, init);
+        this._init = init;
+      }
+      get pointerType() {
+        return this._init.pointerType;
+      }
+      get pageX() {
+        return this._init.pageX;
+      }
+      get pageY() {
+        return this._init.pageY;
+      }
+    };
+  });
+  afterAll(() => {
+    // @ts-ignore
+    delete global.PointerEvent;
+  });
+}
+
+/**
  * Must **not** be called inside an `act` callback!
  *
  * \@testing-library/user-event's `type` helper doesn't call `act` every keystroke.
